+++ WssElixir/app/src/main/java/com/example/wsselixir/view/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.wsselixir
+package com.example.wsselixir.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.wsselixir.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
