+++ WssElixir/app/src/main/res/layout/activity_myinformation.xml
@@ -8,7 +8,7 @@
         android:id="@+id/tvMyInfoName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/all_title_name"
+        android:text="@string/allTitleName"
         android:textSize="30sp"
         android:textStyle="bold"
         android:layout_marginTop="250dp"
@@ -31,7 +31,7 @@
         android:id="@+id/tvMyInfoMBTI"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/all_title_mbti"
+        android:text="@string/allTitleMbti"
         android:textSize="30sp"
         android:textStyle="bold"
         android:layout_marginTop="30dp"
